Implemented a proper representation of a deck.
It is implemented as a sort of doubly linked list. Running 1 million games showed that it performs about 40% faster than the previous implementation of recreating lists all the time. Not that the game is really that performance bound on modern PCs, but why leave performance on the table when a relatively simple data structure can speed things up quite a bit. Lowered the calculation time on the very hard difficulty from 2s to 1.5s because of this. Might even lower it to 1s, but have to test it more before I do that.

Also, while doing this I noticed that there was a bug when placing a Martyr behind a Gladiator where the Gladiator would incorrectly deal ability damage even though the Martyr was the card that was being attacked.