Changed how Guard works.
Guard now reduces the damage it receives by 1, the exact same way as if a Shieldmaiden was behind it. Shieldmaiden obviously still works and would, if placed behind a Guard, mean a damage reduction of 2.

It now seems to be one of the better cards, since a little testing with ai vs ai playouts seems to indicate that the computer now chooses to play Guard during a game about 50% of the time.